+++ Assets/Game Assets/Scripts/Weight.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public class Weight : MonoBehaviour
 {
-    public int weight = 1;
+    public float weight = 1;
 }
